+++ source/main/commands/update-check.js
@@ -16,11 +16,12 @@ const ZettlrCommand = require('./zettlr-command')
 const got = require('got')
 const semver = require('semver')
 const showdown = require('showdown')
+const { app } = require('electron')
 
 const { trans } = require('../../common/lang/i18n.js')
 
 const REPO_URL = require('../../common/data.json').repo_url
-const CUR_VER = require('../../package.json').version
+const CUR_VER = app.getVersion()
 
 class UpdateCheck extends ZettlrCommand {
   constructor (app) {