+++ source/renderer/sidebar/sidebar.vue
@@ -149,7 +149,7 @@
 // Please do not ask me why I have to explicitly use the "default" property
 // of some modules, but not others. The vue-loader is a mess when used with
 // ES6 CommonJS-modules in a exports/require-environment.
-// const tippy = require('tippy.js').default
+const tippy = require('tippy.js').default
 const findObject = require('../../common/util/find-object.js')
 const { trans } = require('../../common/lang/i18n.js')
 const TreeItem = require('./tree-item.vue').default
@@ -476,7 +476,7 @@ module.exports = {
 
       // Create instances for all elements without already existing
       // tippy-instances.
-      window.tippy('#sidebar [data-tippy-content]', {
+      tippy('#sidebar [data-tippy-content]', {
         delay: 100,
         arrow: true,
         duration: 100,