+++ source/renderer/zettlr-body.js
@@ -14,6 +14,12 @@
  *
  * END HEADER
  */
+require('jquery-ui/ui/data')
+require('jquery-ui/ui/scroll-parent')
+require('jquery-ui/ui/version')
+require('jquery-ui/ui/widget')
+require('jquery-ui/ui/widgets/mouse')
+require('jquery-ui/ui/widgets/sortable')
 
 const ZettlrCon = require('./zettlr-context.js')
 const ZettlrNotification = require('./zettlr-notification.js')