+++ source/renderer/zettlr-renderer.js
@@ -330,21 +330,25 @@ class ZettlrRenderer {
     // than in main.
     reconstruct(nData)
     this._paths = nData
-    if (this.getCurrentDir() != null) {
-      this.setCurrentDir(this.getCurrentDir().hash)
-    } else {
-      this.setCurrentDir(null) // Reset
-    }
+
+    // Pass on the new paths object as is to the store.
+    global.store.renewItems(nData)
 
     // Trigger a refresh for all affected places
     this._attachments.refresh()
     this._editor.signalUpdateFileAutocomplete()
 
-    // Pass on the new paths object as is to the store.
-    global.store.renewItems(nData)
-
     // Finally, synchronize the file descriptors in the editor
     this._editor.syncFiles() // DEBUG
+
+    // NOTE: We have to set the directory last because it will re-execute a
+    // potential search, leading to an error if the store, e.g., does not
+    // have the correct list of files.
+    if (this.getCurrentDir() != null) {
+      this.setCurrentDir(this.getCurrentDir().hash)
+    } else {
+      this.setCurrentDir(null) // Reset
+    }
   }
 
   /**